Made server.h to handle sockets in generator program
